Small type correction

+++ src/client/component/getinfo.cpp
@@ -141,8 +141,8 @@ struct component final : generic_component {
       info.set("sub_protocol", std::to_string(SUB_PROTOCOL));
       info.set("playmode", std::to_string(static_cast<int32_t>(
                                game::com::Com_SessionMode_GetMode())));
-      info.set("gamemode",
-               std::to_string(game::com::Com_SessionMode_GetGameMode()));
+      info.set("gamemode", std::to_string(static_cast<int32_t>(
+                               game::com::Com_SessionMode_GetGameMode())));
       info.set("sv_running", std::to_string(game::server_running()));
       info.set("dedicated", game::is_server() ? "1" : "0");
       info.set("hc", std::to_string(game::com::Com_GametypeSettings_GetUInt(
